(Tech) rename api/helpers to avoid mocks conflicts
