+++ src/api/gen/api.ts
@@ -13,7 +13,7 @@
  */
 import url from "url";
 
-import { getAuthenticationHeaders, handleGeneratedApiResponse, safeFetch } from "api/helpers";
+import { getAuthenticationHeaders, handleGeneratedApiResponse, safeFetch } from "api/apiHelpers";
 import { EmptyResponse } from 'libs/fetch'
 
 import { Configuration } from "./configuration";
