+++ src/features/auth/forgottenPassword/ForgottenPassword/tests/ForgottenPassword.native.test.tsx
@@ -3,7 +3,7 @@ import React from 'react'
 import waitForExpect from 'wait-for-expect'
 
 import { navigate, replace } from '__mocks__/@react-navigation/native'
-import { ApiError } from 'api/helpers'
+import { ApiError } from 'api/apiHelpers'
 import { ForgottenPassword } from 'features/auth/forgottenPassword/ForgottenPassword/ForgottenPassword'
 import { MonitoringError } from 'libs/monitoring'
 import { reactQueryProviderHOC } from 'tests/reactQueryProviderHOC'
