+++ src/features/auth/signup/PhoneValidation/SetPhoneValidationCode.tsx
@@ -9,8 +9,8 @@ import { MaskedTextInput } from 'react-native-mask-text'
 import styled from 'styled-components/native'
 
 import { api } from 'api/api'
+import { ApiError, extractApiErrorMessage } from 'api/apiHelpers'
 import { UserProfilingFraudRequest } from 'api/gen'
-import { ApiError, extractApiErrorMessage } from 'api/helpers'
 import { useSendPhoneValidationMutation, useValidatePhoneNumberMutation } from 'features/auth/api'
 import { QuitSignupModal, SignupSteps } from 'features/auth/components/QuitSignupModal'
 import { useAppSettings } from 'features/auth/settings'
