+++ src/features/bookOffer/components/BookingDetails.tsx
@@ -3,8 +3,8 @@ import { useNavigation } from '@react-navigation/native'
 import React, { useEffect } from 'react'
 import styled from 'styled-components/native'
 
+import { isApiError } from 'api/apiHelpers'
 import { OfferStockResponse } from 'api/gen'
-import { isApiError } from 'api/helpers'
 import { UseNavigationType } from 'features/navigation/RootNavigator'
 import { analytics } from 'libs/analytics'
 import { campaignTracker, CampaignEvents } from 'libs/campaign'
