+++ src/features/bookings/components/CancelBookingModal.tsx
@@ -3,8 +3,8 @@ import { useNavigation } from '@react-navigation/native'
 import React, { FunctionComponent } from 'react'
 import styled from 'styled-components/native'
 
+import { extractApiErrorMessage } from 'api/apiHelpers'
 import { UserProfileResponse } from 'api/gen'
-import { extractApiErrorMessage } from 'api/helpers'
 import { Booking } from 'features/bookings/components/types'
 import { useUserProfileInfo } from 'features/home/api'
 import { Credit, useAvailableCredit } from 'features/home/services/useAvailableCredit'
