+++ src/features/offer/api/useOffer.ts
@@ -1,8 +1,8 @@
 import { useQuery } from 'react-query'
 
 import { api } from 'api/api'
+import { ApiError } from 'api/apiHelpers'
 import { OfferResponse } from 'api/gen'
-import { ApiError } from 'api/helpers'
 import { OfferNotFoundError } from 'libs/monitoring'
 import { QueryKeys } from 'libs/queryKeys'
