+++ swagger_codegen/gen_templates/api.mustache
@@ -4,7 +4,7 @@
 
 import url from "url";
 
-import { getAuthenticationHeaders, handleGeneratedApiResponse, safeFetch } from "api/helpers";
+import { getAuthenticationHeaders, handleGeneratedApiResponse, safeFetch } from "api/apiHelpers";
 import { EmptyResponse } from 'libs/fetch'
 
 import { Configuration } from "./configuration";
